refactor: 컴포넌트명 변경
- 컴포넌트의 계층이 더 잘보이도록(컨벤션에 맞게), 컴포넌트명을 변경했다.
